Backend/requirement.txt
flask
flask-sqlalchemy
flask-wtf
mysqlclient
pyjwt
boto3
